moved LocationMarker to its own component
